feat: moved the Trie.ts file into the newly created data-structures directory. The Trie.ts has also been modified to support a prefix validation system as opposed to a lexer.
